HashMapBase: fixed type signature for 'libstdhl::Math::max' calls

+++ src/execute/HashMapBase.h
@@ -249,7 +249,7 @@ class HashMapBase
     explicit HashMapBase( std::size_t initialCapacity )
     : m_buckets( nullptr )
     , m_size( 0UL )
-    , m_capacity( libstdhl::Math::max( initialCapacity, 1 ) )
+    , m_capacity( libstdhl::Math::max( initialCapacity, static_cast< std::size_t >( 1 ) ) )
     , m_lastEntry( nullptr )
 #ifdef HASH_MAP_PERF
     , m_performanceStatistics()
